Add recipe orphan management in UserRecipeService and enhance UserRecipeRepository

+++ src/main/java/com/project/demo/services/UserRecipeService.java
@@ -6,6 +6,7 @@ import com.project.demo.logic.entity.user_recipe.*;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Objects;
@@ -19,15 +20,31 @@ public class UserRecipeService {
     @Autowired
     private RecipeRepository recipeRepository;
 
+    @Autowired
+    private RecipeService recipeService;
 
     public UserRecipe save(Long userId, Long recipeId) {
         UserRecipe userRecipe = new UserRecipe(userId, recipeId);
         return userRecipeRepository.save(userRecipe);
     }
 
+    @Transactional
     public void delete(Long userId, Long recipeId) {
+        boolean exists = userRecipeRepository.findByUserIdAndRecipeId(userId, recipeId).isPresent();
+
+        if (!exists) {
+            throw new RuntimeException("La receta ya no está relacionada con el usuario.");
+        }
+
         userRecipeRepository.deleteByUserIdAndRecipeId(userId, recipeId);
+
+        boolean isOrphan = userRecipeRepository.findByRecipeId(recipeId).isEmpty();
+
+        if (isOrphan) {
+            recipeRepository.deleteById(recipeId);
+        }
     }
+
 
     public List<Recipe> getAllRecipesByUserId(Long userId) {
         List<Long> recipeIds = userRecipeRepository.findByUserId(userId)
